### 2017-04-26 #113 @jirkalewandowski
gitlabr
`gitlabr` provides access to the API of the project and repository management software [gitlab](https://about.gitlab.com/). Files, issues, comments, etc. can be read and modified from R through gitlabr on either your self-hosted installation of gitlab or the public gitlab.com.

```
Package: gitlabr
Title: Access to the Gitlab API
Version: 0.9
Authors@R: person("Jirka", "Lewandowski", email = "jirka.lewandowski@wzb.eu", role = c("aut", "cre"))
Description: Provides R functions to access the API of the project and
    repository management web application gitlab. For many common tasks (repository
    file access, issue assignment and status, commenting) convenience wrappers
    are provided, and in addition the full API can be used by specifying request
    locations. Gitlab is open-source software and can be self-hosted or used on
    gitlab.com.
    R (>= 3.1.2),
    magrittr
    base64enc,
    purrr (>= 0.2.2),
    tibble (>= 1.1),
    yaml
    devtools (>= 1.8.0),
    roxygen2,
    R.rsp,
    shiny (>= 0.13.0)
VignetteBuilder: R.rsp
```

https://github.com/jirkalewandowski/gitlabr

People who use R in a team that uses gitlab for code and/or data hosting and/or project management.

There is [gh](https://github.com/r-pkgs/gh) providing similar things for Github.com, but Gitlab has a free & open source self-hosting variant and is used internally by many organizations, including academics.

- [x] Do you intend for this package to go on CRAN?  *It already is on CRAN.*

- [x] Does the package conform to [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? Please describe any exceptions: *The package does not use covr, since the test suite seems to fail in the environment covr creates. I'm very happy to discuss this and/or get help with fixing. See the branch [codecov](https://github.com/jirkalewandowski/gitlabr/tree/codecov)*
