### 2017-04-28 #113 @noamross
Thank you for the submission, @jirkalewandowski!  This package fits under our reproducibility category.  A few small things before we assign reviewers:

1) We generally move accepted packages into the rOpenSci organization upon acceptance. We do this for several reasons, primarily so that we can do testing, serve rOpenSci packages through our API, and also so we can chip in maintenance if authors fall behind. Our few exceptions have been when the package is already under the umbrella of another group, such as Bioconductor, that serves the same role. If accepted, we'd like to set up a way to at least mirror to the rOpenSci organization.

2) We would strongly prefer that the package be hosted in its installable state, meaning that it includes NAMESPACE, documentation, etc., even if those are generated by *roxygen2*.

3) Could you provide some developer-aimed documentation on the minimum setup needed to run the tests locally such as env variables and acquiring appropriate tokens? (In a `CONTRIBUTING.md` file would be appropriate). I plan to have both an experienced and a new gitlab user as reviewers (I myself am the latter), and this will also make it easier gitlab novices assist with the *covr* issue and run our other checks.
