### 2017-05-02 #113 @jirkalewandowski
Dear @noamross, thank you for taking care of this submission and your remarks.

ad 1. Sounds great, we could move the jirkalewandowski/gitlabr github project to the ropensci organization then. Since the github repo is a mirror of the development repo on https://gitlab.points-of-interest.cc anyways, there is no need for the jirkalewandowski/gitlabr repo in addition to ropensci.

ad 2. & 3. Sounds also great. I'll take of both issues within the next two weeks and post back here.
