### 2017-05-12 #113 @jirkalewandowski
Dear @noamross, I added the roxygen-generated files to the github repository and a CONTRIBUTING.md with the requested information.
