### 2017-05-12 #113 @noamross
Thank you @jirkalewandowski!

I've run goodpractice::gp() (skipping **covr** checks).  Here are the results.  None of them are major/mandatory but should be considered guidance for yourself and the reviewers.

── GP gitlabr ───────────────────────────

It is good practice to

  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes, and poor interaction with other packages. Use "Imports" instead.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines shorter than 80 characters

    R/ci.R:12:1
    R/ci.R:26:1
    R/ci.R:33:1
    R/ci.R:36:1
    R/ci.R:41:1
    ... and 101 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using vapply() instead.

    R/ci.R:38:41

  ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import only the specific functions you need.
──────────────────────────────────────────────
`devtools::spell_check()` yielded to real hits:
