### 2017-05-12 #113 @noamross
@rmflight @jennybc As part of your review please look into the *covr* issue. If it's not clear to you what the issue is we can call in some additional expertise.
