### 2017-07-11 #113 @jennybc
---

Overall, I'm excited by this package and think it has a really interesting design. It provides both high- and low-level functions for interacting with the gitlab API. It's largely well-constructed, with idiomatic R code and very thorough documentation and explanation of the important design features.

I was impressed that I could run the test suite with a modest amount of tweaking and knowledge. However my main concern right now is that I can't do simple things with the gitlab.com API. I haven't dug deep enough to tell if there's a problem with gitlabr itself or with its documentation, which is causing user error. I can make basic requests using "raw" httr, so it doesn't appear to be a fatal problem with my gitlab.com account. I note there are no examples or tests against gitlab.com, which feels like an oversight. I think functionality and/or documentation has gotten over-tuned against the author's gitlab instance.

**A statement of need**

This is not *explicitly* done in the README. However, this is definitely present in other places. It would be good to add 1-2 sentences to README.

**Vignette(s)**

I have run the vignette code that is not `eval = FALSE`, after substituting my own username and password. My main concern is that I cannot run equivalent code against gitlab.com (more below).

**Examples**

Many functions lack examples. I wouldn't expect universal example coverage, in any event, but more examples would be good.

**Functionality:**

I have not been able to run basic code against gitlab.com. I *can* run essentially all code I see in tests, vignette, etc. against gitlab.points-of-interest.cc. It's not clear if this is a real problem or a documentation problem, leading to user error. Either way, it needs to be addressed.

It would probably be best if examples, README, etc. featured gitlab.com, as most users would probably be doing that. It might also reveal now or in the future if there is some underlying problem that I am experiencing.

For example, here's code adapted from the examples in `gl_connection()`:

```r
my_gitlab <- gl_connection(
  "https://gitlab.com",
  private_token = Sys.getenv("GITLAB_PVT_TOKEN")
)
my_gitlab("projects")
```

First, the `gl_connection()` call needs to specify the name of the second argument or else it's interpreted as the `login`. Second, the `my_gitlab()` call hangs indefinitely for me.

**Automated tests:**

I was able to run tests for this package without too much pain, which is pretty impressive, so big kudos for that! I have yet to write tests for an API-wrapping package that achieves this worthy goal. Some thoughts and suggestions:

  * Why test against your gitlab server vs gitlab.com? It's not completely hard-wired, given the URL in `environment.yml` but I suspect there would be some undocumented setup required if I actually changed the target gitlab server.
  * The instructions don't tell me how to populate `environment.yml`. Where do I get this info on gitlab? Is the token supposed to be the private token displayed in the account tab of user settings or an access token generated from the access tokens tab?
  * Once I populate `environment.yml` with real info, I should `.gitignore` it, right? That would be good to clarify. If testing on CI, then I assume there's an encryption strategy for this file. Or maybe secure env vars are used?
  * Each test file should use `context("blah blah")` at the top for nicer reporting. I've done this is my review branch.
  * I had to `skip()` the test about issue editing, because I don't have necessary permission. Would be good to implement this such that tester is editing an issue they have permission on. Again, pretty impressive that it's the *only* test I had to skip!  I've done this is my review branch.
  * [These lines re: API version in `testthat.R`](https://github.com/jirkalewandowski/gitlabr/blob/3525479b1316abd2f2122f7778be2c3c4f8b3d06/tests/testthat.R#L4-L14) are pretty unusual. I can't infer the exact goal, but wonder if there's a better / more conventional way to achieve it. Obviously, I understand it's got something to do with testing against different API versions.
  * I've moved common env var setup into `helper.R` and out of individual test files. Looking at the top of these files, I suspect even more setup could probably be similarly centralized? `helper.R` is always sourced by `devtools::load_all()` so this makes for a nice workflow even when developing / running tests interactively, i.e. the info doesn't *need* to live in each file.

**Packaging guidelines**:

I recommend use of `README.Rmd`. I find this makes it much harder for `README.md` to slowly drift out of sync with what the package actually does.

I needed to explicitly qualify `purrr::set_names()` to pass `R CMD check`.

As mentioned above, many functions do not have an example.

Package dependencies. Re: the pipe `%>%` and friends: I advise to move magrittr from "Depends" to "Imports". You can then import individual operators that are used. In my review branch, I've done this.

#### Other notes

**dplyr dependency**: I would seriously consider NOT depending on dplyr. I realize this might be highly impractical. But if dplyr can be removed with modest effort, it might be worth it. dplyr is more end-user facing and not optimized for use within other packages, whereas tibble and purrr are designed with this usage in mind.

**rprojroot**: [rprojroot](https://cran.r-project.org/web/packages/rprojroot/index.html) might be useful for hardening path formation, e.g., below `tests/`. Or even testthat offers some support for same. Paths like `yaml::yaml.load_file("../../.gitlab-ci.yml")` have a way of breaking. However, I acknowledge the current strategy must be working pretty well, since gitlabr is pretty mature.

**Deprecated functions**: Currently the deprecated functions dominate the function index. I wonder if ... you could `@keyword` these `internal` in order to keep them around but suppress their inclusion here?

**Tibbles of results**: For `gl_list_projects()` (and perhaps others): could you hoist `name` and `name_with_namespace` to the front for human readability?

#### Things I did not check

I did not have time to look into the covr issue.
