### 2017-07-11 #113 @jirkalewandowski
Dear @jennybc @rmflight -- thank you very much for the detailed reviews! I learned several things already from reading and will start implementing the changes you suggest over the next months, as I see how this will improve the package in many respects.
