### 2017-07-11 #113 @jennybc
I thought more about my experience and what I learned yesterday about the differences between gitlab and GitHub: gitlab lets you see all public projects and regards them as "yours" when you list. Coupled with automatic pagination, this is why `gl_list_projects()` hangs indefinitely for me on gitlab.com. Who knows how many projects its trying to retrieve?

If I limit to one page, I do succeed:

``` r
library(gitlabr)
my_gitlab <- gl_connection(
  gitlab_url = "https://gitlab.com",
  private_token = Sys.getenv("GITLAB_PVT_TOKEN")
)
my_gitlab(gl_list_projects, page = 1)
#> # A tibble: 20 x 50
#>         id
#>      <chr>
#>  1 3683998
#>  2 3683931
#>  3 3683891
#>  4 3683865
#>  5 3683855
#>  6 3683853
#>  7 3683770
#>  8 3683767
#>  9 3683766
#> 10 3683753
#> 11 3683719
#> 12 3683693
#> 13 3683669
#> 14 3683665
#> 15 3683664
#> 16 3683657
#> 17 3683651
#> 18 3683640
#> 19 3683628
#> 20 3683617
#> # ... with 49 more variables: description <chr>, archived <chr>,
#> #   visibility <chr>, ssh_url_to_repo <chr>, http_url_to_repo <chr>,
#> #   web_url <chr>, owner.name <chr>, owner.username <chr>, owner.id <chr>,
#> #   owner.state <chr>, owner.avatar_url <chr>, owner.web_url <chr>,
#> #   name <chr>, name_with_namespace <chr>, path <chr>,
#> #   path_with_namespace <chr>, container_registry_enabled <chr>,
#> #   issues_enabled <chr>, merge_requests_enabled <chr>,
#> #   wiki_enabled <chr>, jobs_enabled <chr>, snippets_enabled <chr>,
#> #   created_at <chr>, last_activity_at <chr>,
#> #   shared_runners_enabled <chr>, lfs_enabled <chr>, creator_id <chr>,
#> #   namespace.id <chr>, namespace.name <chr>, namespace.path <chr>,
#> #   namespace.kind <chr>, namespace.full_path <chr>, import_status <chr>,
#> #   star_count <chr>, forks_count <chr>, open_issues_count <chr>,
#> #   public_jobs <chr>, only_allow_merge_if_pipeline_succeeds <chr>,
#> #   request_access_enabled <chr>,
#> #   only_allow_merge_if_all_discussions_are_resolved <chr>,
#> #   approvals_before_merge <chr>, default_branch <chr>,
#> #   forked_from_project.id <chr>,
#> #   forked_from_project.http_url_to_repo <chr>,
#> #   forked_from_project.web_url <chr>, forked_from_project.name <chr>,
#> #   forked_from_project.name_with_namespace <chr>,
#> #   forked_from_project.path <chr>,
#> #   forked_from_project.path_with_namespace <chr>
```

So I think this (at least) is a documentation issue, but a pretty important one. Perhaps there should be a page limit of last resort, as well, to prevent an attempt to retrieve all public projects on gitlab, for example.
