### 2017-09-18 #113 @jirkalewandowski
Dear @noamross, thanks for the reminders and sorry for the late response, I've been offline for a while. My plan is to incorporate the feedback in October, most likely in combination with packing together gitlabr 1.0, the first version that will no longer support the old API v3, but only v3.
