### 2020-08-12 #113 @statnmap
I'd like to help in finishing this process. Indeed, the maintainer passed away last year, so I decided to take over the development of this package but I am not sure to find time and resources to do it alone.
I did not had the opportunity to know Jirka. I am just a recent user of his package.

I already cloned the original Github repository and started accounting for @jennybc comments: https://github.com/statnmap/gitlabr/
I cloned it in a way to allow issues on the repository, there is no link to the original repository.

I think the Readme is now closer to a direct reproducible example, provided that the user has a token.
Now there is obviously some needs in the vignette, functions examples and unit tests as also pointed by @rmflight .
And there is a need to upgrade according to last {tibble} and {dplyr} versions. I wrote a fix, but did not went deep into it.

I will try to continue the process, but it is difficult to find some time on it.
Can I find some help among rOpenSci developers? Could you use your network to ask for collaboration?
