### 2020-08-12 #113 @noamross
Hello @statnmap.  We're saddened to hear of the maintainer's passing.  We are encouraged to hear you are interested in updating and maintaining the package and would be glad to help find you some support.   Can you provide an email so we can invite you to our Slack workspace where we'll ping for volunteers?
