### 2020-08-14 #113 @statnmap
I joined the Slack workspace. Thanks.
I take my 3 weeks vacation and when I come back, we'll ping for volunteers.
I'll have to open some issues first to define what is to be done.
